ete-digital/backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Security
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.0
python-multipart==0.0.6

# Redis
redis==5.0.1
hiredis==2.3.2

# HTTP Client
httpx==0.26.0

# Email
email-validator==2.3.0

# Object Storage
minio==7.2.3

# Background Tasks (Celery alternative for Python)
celery==5.3.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Monitoring
sentry-sdk[fastapi]==1.40.0
